Changed jumping mechanics.
Originally, only one boolean variable, 'onGround', was used for handling
jumping. When considering platforms, this meant that the player could
jump anytime after landing on a platform (including after sliding off
it) instead of only being able to jump when on top of it. I added a
second boolean variable 'jumping' to track time after the jump key has
been pressed. It may be possible to replicate the appropriate behavior
with only one boolean variable, but two is fine for now.