Removed chance of platforms intersecting.
Platforms should no longer spawn in contact with eachother. Everytime a
platform is supposed to be spawned, first a random x-coordinate is
chosen and then a random y-coordinate that is not already occupied by a
platform is chosen. This seems to have improved platform distribution
(possibly just as a side effect of the particular implementation).

Currently, platforms that, based on the chosen x-coordinate, cannot
spawn at any y-coordinate will instead be spawned above the canvas.