Switched to using gameTimer for spawning objects.
This switch removes the need for separate spawnCounter and
platformCounter variables. This does introduce some unexpected behavior
for spawns (and probably platforms too) where multiple will be spawned
in more rapid than usual succession (e.g., two at once). This seems to
be because of concurrent changes to spawnCounterTarget. For now, I don't
consider it a bug.