Added jetpack powerup.
The player can now press and hold the jump button in order to hover in
air after collecting a green square that falls from the sky. The powerup
lasts for 10 seconds.

It might be valuable to dedicate some time to reconsidering or
abstracting parts of the codebase that were involved in adding this
powerup. While it wasn't too bad, there were a lot of different pieces
that needed to be included across multiple files in order to add powerup
functionality, and it would be nice if that process was more
comfortable.