Remove raf variable
This variable, used to store the return value of the call to the
window.requestAnimationFrame function, would typically be used as an
argument to the window.cancelAnimationFrame function. Since there
doesn't seem to be a need for this function in the game, I think the
variable is a bit of copy-pasted code that can be safely removed.