Add double points powerup
There is a good amount of copy-pasted/redundant code that went into this
addition. It would be valuable to invest time in generalizing powerups
so that they can be added more easily, as well as generalizing the
process by which powerups are randomly selected for spawning in-game.

Currently, if both a jetpack and a double points powerup are active at
the same time, the countdown timers will display on top of eachother.