Increase x-spacing between two intial platforms
Originally the gap was deceptively sized such that (if the two platforms
had the same y-values) the player could walk across, even though it did
not appear that the player model could. While this may indicate separate
issues with player-platform collision detection, increasing the spacing
between the two platforms is a basic solution.