Tweak approach to getting random platform y-values
The current approach is to get a random number based on a random value
between 0 and 1 squared, which emphasizes low y-values. This leads to
the majority of platforms being distributed high on the screen and in a
line, with platforms rarely sharing a vertical axis. This change shifts
to using a random value between 0 and 1, un-squared, which gives an
apparently fun-er and random-er distribution of platforms.