Add pause button
With this change, the game can be paused/unpaused by pressing spacebar.
When the game is paused, a large red "PAUSED" appears in the middle of
the screen.

Because of the way the keydown event listener handles toggling the pause
variable, I believe the "PAUSED" is drawn for an additional frame when
moving from the paused state to the unpaused state, and so technically
an extra frame than it needs to be. Since it doesn't seem to be worth
the mess to avoid this extra frame, I've left it be.