Add possibility of spawns wobbling in air
This change is another attempt to increase the variety of ways in which
spawns can behave before hitting the platform or the ground. While the
wobbling behavior introduced with this change feels fairly erratic (in a
good way), it might be a valuable to abstract/parametrize additional
aspects of the wobbling code in the future to allow more control/more
randomization.