Improve behavior of music play/pause button
There's an issue where the "p: [play/pause] music" instruction flickers
while an audio track is loading (as the page loads for the first time or
while cycling songs) as it shifts from the word "pause" to the word
"play". This change remedies this issue by creating an audioMessage
variable and making changes to the behavior of the 'p' button such that
it provides a filtered indicator for the state of the audio track,
something more appropriate than just 'audio.paused'.