Add possibility of spawn tracking player movement
Current implementation is pretty rudimentary: the spawns that
are set to track have very slow y-movement and adjust the direction of
their x-movement based on the x-position of the player.

Adding more spawn variations is getting increasingly messy, and so it
may be a good idea to reconsider the spawn class structure in the near
future.