Display powerup timers on different lines
Currently powerup timers on the play screen will draw on top of each
other if both powerups are active at the same time. This change provides
a (likely temporary) fix to this problem by always drawing the text for
the doublepoints powerup timer lower than the text for the jetpack
powerup timer. Some other screen instructions were moved lower on the
screen in order to better accommodate this shift.