Fix lag by drawing text to separate canvas
Evidently, the stroking introduced in 0a3da5b came with a heavy hit to
the performance of the game. In this change, the game screen text that
does not update as the game goes on (i.e., most of the instruction text)
is instead drawn on game (re)start to a textCanvas object that is then
copied to the main canvas during the draw call. This significantly
speeds up the game.