Remove/tweak variable definitions in setup.js
The setup.js is intended to define a number of global variables. The use
of the 'var' keyword when initializing variables outside the setup
function is unecessary, since those variables are already at the global
scope. Further, defining (but not initializing) variables above the
setup function is largely unecessary because variables are already
initialized into the global scope by not using the 'var' keyword.