Add fast falling mechanic
With this change, when the player is in air and presses the down key,
they start speeding up at a multiple of the usual gravity, and can
exceed terminal velocity as long as they hold down.  When they release
the down key, they return to terminal velocity if they are exceeding it.