Expand affect that freeze has on spawns
With this change, multiple forms of spawn movement are affected by the
freeze powerup, not just the spawn gravity. In addition, spawns will not
spawn while the freeze powerup is active; this is to prevent a massive
wall of spawns from collecting on the upper part of the screen and
crashing down on the player when the freeze powerup ends.